// src/features/orders/types/order-data.ts
import { OrderDetail } from "./order-detail";

interface OrderData {
  id: number;
  number: string;
  class: null;
  space_type: string;
  space_id: number;
  model_type: string;
  model_id: null;
  type_type: null;
  type_id: null;
  input_type: null;
  input_id: null;
  output_type: null;
  output_id: null;
  sender_type: string;
  sender_id: number;
  receiver_type: string;
  receiver_id: null;
  handler_type: string;
  handler_id: string;
  input_address: null;
  output_address: null;
  request_time: null;
  sent_time: string;
  received_time: null;
  handler_number: null;
  total: number;
  fee: string;
  fee_rules: null;
  sender_notes: null;
  receiver_notes: null;
  handler_notes: null;
  status: string;
  notes: null;
  created_at: string;
  updated_at: string;
  deleted_at: null;
  sent_date: null;
  description: null;
  relation_type: null;
  relation_id: null;
  files: unknown[];
  parent_type: null;
  parent_id: null;
  total_details: number;
  tags: unknown[];
  links: unknown[];
  details: OrderDetail[];
}

export type { OrderData };